Paart 4 Exercises 4.21 - 4.22

+++ controllers/blogs.js
@@ -53,7 +53,20 @@ blogRouter.delete("/:id", async (request, response) => {
     return response.status(400).json({ error: "id is required" })
   }
 
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: "not authorized" })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
+
   response.status(204).end()
 })
